Code cleanup

+++ libraries/chip8++/system/InstructionDecoder.hpp
@@ -62,7 +62,7 @@ namespace CHIP8
 				for (int i = 0; i < 0x1000; i++)
 				{
 					auto num = (InstructionInfo.opcode & InstructionInfo.mask) | (i & ~InstructionInfo.mask);
-					InstructionTable[num] = instruction;
+					InstructionTable[size_t(num)] = instruction;
 				}
 			}
 			else
